Move overplot menu to viewer.
